+++ chain/network/src/raw/mod.rs
@@ -1,6 +1,6 @@
 mod connection;
 
-pub use connection::{ConnectError, Connection, ReceivedMessage};
+pub use connection::{ConnectError, Connection, DirectMessage, Message, RoutedMessage};
 
 #[cfg(test)]
 mod tests;