+++ core/store/src/db/refcount.rs
@@ -65,6 +65,27 @@ pub(crate) fn strip_refcount(mut bytes: Vec<u8>) -> Option<Vec<u8>> {
     None
 }
 
+/// Sets the refcount to the given value.
+///
+/// This method assumes that the data already contains a reference count stored
+/// in the last 8 bytes. It overwrites this value with the new value.
+///
+/// Returns None if the input bytes are too short to contain a refcount.
+pub(crate) fn set_refcount(data: &mut Vec<u8>, refcount: i64) -> io::Result<()> {
+    const BYTE_COUNT: usize = std::mem::size_of::<i64>();
+
+    if let Some(len) = data.len().checked_sub(BYTE_COUNT) {
+        let refcount: [u8; BYTE_COUNT] = refcount.to_le_bytes();
+        data[len..].copy_from_slice(&refcount);
+        Ok(())
+    } else {
+        Err(io::Error::new(
+            std::io::ErrorKind::InvalidData,
+            "The data is too short to set refcount",
+        ))
+    }
+}
+
 /// Encode a positive reference count into the value.
 pub(crate) fn add_positive_refcount(data: &[u8], rc: std::num::NonZeroU32) -> Vec<u8> {
     [data, &i64::from(rc.get()).to_le_bytes()].concat()
@@ -211,6 +232,19 @@ mod test {
 
         let rc = std::num::NonZeroU32::new(2).unwrap();
         assert_eq!(MINUS_TWO, &super::encode_negative_refcount(rc));
+    }
+
+    #[test]
+    fn set_refcount() {
+        fn test(want: Vec<u8>, data: &[u8], refcount: i64) {
+            let mut data = data.to_vec();
+            let result = super::set_refcount(&mut data, refcount);
+            assert!(result.is_ok());
+            assert_eq!(want, data);
+        }
+
+        test(PLUS_TWO.to_vec(), b"\0\0\0\0\0\0\0\0", 2);
+        test(MINUS_TWO.to_vec(), b"\0\0\0\0\0\0\0\0", -2);
     }
 
     #[test]